+++ src/runtime/mod.rs
@@ -1,5 +1,7 @@
 mod execution;
 mod persistence;
+mod step;
 
-pub use execution::{change_wd, execute, run_command, Outcome, Step};
+pub use execution::{change_wd, execute, run_command, Outcome};
 pub use persistence::{forget, load, persist};
+pub use step::Step;
